+++ india_compliance/gst_india/overrides/test_advance_payment_entry.py
@@ -7,7 +7,7 @@ from frappe.tests.utils import FrappeTestCase
 from erpnext.accounts.doctype.payment_entry.payment_entry import (
     get_outstanding_reference_documents,
 )
-from erpnext.accounts.doctype.unreconcile_payments.unreconcile_payments import (
+from erpnext.accounts.doctype.unreconcile_payment.unreconcile_payment import (
     create_unreconcile_doc_for_selection,
 )
 from erpnext.controllers.stock_controller import show_accounting_ledger_preview
